move Request classes to console module
