+++ eventmesh-dashboard-core/src/main/java/org/apache/eventmesh/dashboard/core/meta/EtcdConnectionService.java
@@ -15,9 +15,8 @@
  * limitations under the License.
  */
 
-package org.apache.eventmesh.dashboard.core.service.meta;
+package org.apache.eventmesh.dashboard.core.meta;
 
-import org.apache.eventmesh.dashboard.core.config.AdminProperties;
 import org.apache.eventmesh.dashboard.service.meta.ConnectionService;
 
 import org.springframework.stereotype.Service;
@@ -26,9 +25,6 @@ import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
 @Service
-public class NacosConnectionService implements ConnectionService {
-
-    public NacosConnectionService(AdminProperties adminProperties) {
-    }
+public class EtcdConnectionService implements ConnectionService {
 
 }
